Make it possible to add addition artifacts that are added to extensions without them being in the project. This is to deal with JDK11 + OpenJFX 11+ on Windows. Because the extension.zip files are BUILT on Linux, this means that they get the javafx-*-linux classified artifacts.
The old client worked because *Oracle* Java 8 has JavaFX included.

JavaFX jars work differently from SWT ones, in that 1) they use a classifier for each Jar and 2) each Jar appears to be for the OS, not the platform. Being as everything is effectively 64 bit now, this is not a massive deal.

We need to exclude this from the root of the zip, and add them to the appropriate OS specific folder.

This way JavaFX will work and we can treat it as updateable like SWT instead of relying on an outdated runtime.